+++ spec/client_spec.rb
@@ -36,6 +36,26 @@ describe 'mcollective::client' do
           .with_content(/plugin.activemq.*#{word}/)
       end
     end
+
+    context 'with client middleware credentials' do
+      let(:chef_run) {
+        chef_run = ChefSpec::Runner.new(:platform => 'redhat', :version => '6.3')
+        chef_run.node.set['mcollective']['connector'] = 'activemq'
+        chef_run.node.set['mcollective']['stomp']['username'] = 'testuser'
+        chef_run.node.set['mcollective']['stomp']['password'] = 'testpass'
+        chef_run.node.set['mcollective']['stomp']['client_username'] = 'clientuser'
+        chef_run.node.set['mcollective']['stomp']['client_password'] = 'clientpass'
+        chef_run.converge(described_recipe)
+      }
+
+      it 'writes the activemq credentials in the client config' do
+        words = %w{clientuser clientpass}
+        words.each do |word|
+          expect(chef_run).to render_file('/etc/mcollective/client.cfg')
+            .with_content(/plugin.activemq.*#{word}/)
+        end
+      end
+    end
   end
 
   context 'configured to use rabbitmq' do
@@ -58,7 +78,27 @@ describe 'mcollective::client' do
           .with_content(/plugin.rabbitmq.*#{word}/)
       end
     end
-end
+
+    context 'with client middleware credentials' do
+      let(:chef_run) {
+        chef_run = ChefSpec::Runner.new(:platform => 'redhat', :version => '6.3')
+        chef_run.node.set['mcollective']['connector'] = 'rabbitmq'
+        chef_run.node.set['mcollective']['stomp']['username'] = 'testuser'
+        chef_run.node.set['mcollective']['stomp']['password'] = 'testpass'
+        chef_run.node.set['mcollective']['stomp']['client_username'] = 'clientuser'
+        chef_run.node.set['mcollective']['stomp']['client_password'] = 'clientpass'
+        chef_run.converge(described_recipe)
+      }
+
+      it 'writes the activemq credentials in the client config' do
+        words = %w{clientuser clientpass}
+        words.each do |word|
+          expect(chef_run).to render_file('/etc/mcollective/client.cfg')
+            .with_content(/plugin.rabbitmq.*#{word}/)
+        end
+      end
+    end
+  end
 
   context 'configured to use redis' do
     let(:chef_run) {